Update up next styling for podcasts
