+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/MainActivity.kt
@@ -76,6 +76,7 @@ private fun WearApp() {
                 navController = navController,
                 pagerState = pagerState,
                 swipeToDismissState = swipeToDismissState,
+                scrollableScaffoldContext = it,
             ) {
                 WatchListScreen(
                     scrollState = it.scrollableState,
@@ -110,6 +111,7 @@ private fun WearApp() {
             NowPlayingPager(
                 navController = navController,
                 swipeToDismissState = swipeToDismissState,
+                scrollableScaffoldContext = it,
             ) {
                 PodcastsScreen(
                     listState = it.scrollableState,
@@ -163,6 +165,7 @@ private fun WearApp() {
             NowPlayingPager(
                 navController = navController,
                 swipeToDismissState = swipeToDismissState,
+                scrollableScaffoldContext = it,
             ) {
                 DownloadsScreen(
                     columnState = it.columnState,
