+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/UpNextScreen.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rxjava2.subscribeAsState
@@ -14,18 +13,16 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.wear.compose.foundation.lazy.ScalingLazyColumn
 import androidx.wear.compose.foundation.lazy.ScalingLazyListState
 import androidx.wear.compose.foundation.lazy.items
-import androidx.wear.compose.material.Chip
-import androidx.wear.compose.material.ChipDefaults
 import androidx.wear.compose.material.MaterialTheme
 import androidx.wear.compose.material.Text
-import au.com.shiftyjelly.pocketcasts.compose.components.PodcastImage
 import au.com.shiftyjelly.pocketcasts.repositories.playback.UpNextQueue
+import au.com.shiftyjelly.pocketcasts.wear.ui.component.EpisodeChip
+import au.com.shiftyjelly.pocketcasts.wear.ui.component.ScreenHeaderChip
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 object UpNextScreen {
@@ -68,26 +65,18 @@ fun UpNextScreen(
         is UpNextQueue.State.Loaded -> {
             val list = (queueState as UpNextQueue.State.Loaded).queue
             ScalingLazyColumn(
-                modifier = modifier.fillMaxWidth(),
                 state = listState,
+                modifier = modifier.fillMaxWidth(),
             ) {
+
+                item { ScreenHeaderChip(LR.string.up_next) }
+
                 items(list) { episode ->
-                    Chip(
+                    EpisodeChip(
+                        episode = episode,
                         onClick = {
                             navigateToEpisode(episode.uuid)
                         },
-                        colors = ChipDefaults.secondaryChipColors(),
-                        label = {
-                            Text(episode.title, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                        },
-                        icon = {
-                            PodcastImage(
-                                uuid = episode.uuid,
-                                dropShadow = false,
-                                modifier = Modifier.size(30.dp)
-                            )
-                        },
-                        modifier = modifier.fillMaxWidth()
                     )
                 }
             }
